feat: pss-cln-pses-contract-roots-002 - Fold excess root implementation tests into private destinations
