+++ pkg/services/provider_sessions/legacy_packages_disposition_test.go
@@ -2,6 +2,7 @@ package providersessions_test
 
 import (
 	"fmt"
+	"os"
 	"path/filepath"
 	"strings"
 	"testing"
@@ -88,21 +89,23 @@ var providerSessionsProductionPublicSurfacePackages = []string{
 	"github.com/portpowered/infinite-you/pkg/services/provider_sessions/transports/http",
 }
 
-// providerSessionsRootBehaviorProofFiles are committed fold-target implementation
-// tests that exercise wire-constructed Details, Inspect, and Project on Codex- and
+// providerSessionsInternalBehaviorProofFiles are folded implementation tests
+// that exercise wire-constructed Details, Inspect, and Project on Codex- and
 // Cursor-backed reader fixtures for the published Service peer surface.
-var providerSessionsRootBehaviorProofFiles = []string{
+var providerSessionsInternalBehaviorProofFiles = []string{
 	"wire_behavioral_proof_test.go",
 	"details_providers_boundary_test.go",
 	"inspect_providers_boundary_test.go",
 	"project_providers_boundary_test.go",
 	"service_test.go",
+	"readers_providers_boundary_test.go",
 }
 
 // TestProviderSessionsRootBehaviorPreserved seals pss-cln-pses-legacy-packages-004:
 // focused wire-constructed behavioral proofs and Providers-root boundary tests
-// remain inventoried and the root-go inventory gate stays green so legacy-sibling
-// cleanup cannot silently drop Details/Inspect/Project observability coverage.
+// remain under provider_sessions/internal and the root-go inventory gate stays
+// green so legacy-sibling cleanup cannot silently drop Details/Inspect/Project
+// observability coverage.
 func TestProviderSessionsRootBehaviorPreserved(t *testing.T) {
 	t.Parallel()
 
@@ -111,31 +114,11 @@ func TestProviderSessionsRootBehaviorPreserved(t *testing.T) {
 		t.Fatalf("VerifyProviderSessionsRootGoInventory() error = %v", err)
 	}
 
-	inventory, err := ownershipinventory.LoadProviderSessionsRootGoInventory(root)
-	if err != nil {
-		t.Fatalf("LoadProviderSessionsRootGoInventory() error = %v", err)
-	}
-
-	inventoried := make(map[string]ownershipinventory.ProviderSessionsRootGoFile, len(inventory.Files))
-	for _, file := range inventory.Files {
-		inventoried[file.File] = file
-	}
-
-	for _, name := range providerSessionsRootBehaviorProofFiles {
-		file, ok := inventoried[name]
-		if !ok {
-			t.Fatalf("root behavior proof file %q missing from committed root-go inventory", name)
-		}
-		if file.Classification != ownershipinventory.ProviderSessionsRootGoFoldTargetImplTest {
-			t.Fatalf(
-				"%s classification = %q, want %q",
-				name,
-				file.Classification,
-				ownershipinventory.ProviderSessionsRootGoFoldTargetImplTest,
-			)
-		}
-		if file.FoldDestination != "pkg/services/provider_sessions/internal" {
-			t.Fatalf("%s foldDestination = %q, want pkg/services/provider_sessions/internal", name, file.FoldDestination)
+	internalRoot := filepath.Join(root, "pkg", "services", "provider_sessions", "internal")
+	for _, name := range providerSessionsInternalBehaviorProofFiles {
+		path := filepath.Join(internalRoot, name)
+		if _, err := os.Stat(path); err != nil {
+			t.Fatalf("internal behavior proof file %q missing: %v", name, err)
 		}
 	}
 }
